refactor(runtime): split js_typed_array_view into typedarray_view module (file-size gate)
