+++ crates/perry-runtime/src/typedarray.rs
@@ -191,7 +191,7 @@ fn data_ptr(ta: *const TypedArrayHeader) -> *const u8 {
 }
 
 #[inline]
-fn data_ptr_mut(ta: *mut TypedArrayHeader) -> *mut u8 {
+pub(crate) fn data_ptr_mut(ta: *mut TypedArrayHeader) -> *mut u8 {
     unsafe {
         if crate::native_arena::is_native_typed_view(ta as *const TypedArrayHeader) {
             crate::native_arena::native_view_data_ptr_mut(ta)
@@ -287,7 +287,7 @@ fn throw_type_error(message: &[u8]) -> ! {
 }
 
 #[cold]
-fn throw_range_error(message: &[u8]) -> ! {
+pub(crate) fn throw_range_error(message: &[u8]) -> ! {
     let msg = crate::string::js_string_from_bytes(message.as_ptr(), message.len() as u32);
     let err = crate::error::js_rangeerror_new(msg);
     crate::exception::js_throw(crate::value::js_nanbox_pointer(err as i64))
@@ -761,111 +761,6 @@ pub extern "C" fn js_typed_array_new(kind: i32, val: f64) -> *mut TypedArrayHead
     }
     // Undefined / null / bool / string → empty typed array.
     typed_array_alloc(kind as u8, 0)
-}
-
-/// `ToIndex(value)` for a typed-array view's `byteOffset` / `length`
-/// arguments (#4103): `undefined` / `NaN` → 0, otherwise `ToIntegerOrInfinity`
-/// truncated toward zero, with a `RangeError` for a negative or
-/// out-of-`[[0, 2^53-1]]` result.
-fn typed_array_view_to_index(value: f64) -> i64 {
-    let jv = crate::value::JSValue::from_bits(value.to_bits());
-    if jv.is_undefined() {
-        return 0;
-    }
-    let n = jv.to_number();
-    if n.is_nan() {
-        return 0;
-    }
-    if n < 0.0 || n > 9_007_199_254_740_991.0 {
-        throw_range_error(b"Invalid typed array length");
-    }
-    n.trunc() as i64
-}
-
-/// `new TA(buffer, byteOffset, length?)` for the non-`Uint8Array` typed-array
-/// kinds — create a view over an `ArrayBuffer` with the spec's offset/length
-/// validation (#4103). Perry still models the result as an owning
-/// `TypedArrayHeader` (the `byteOffset` getter reports 0), but the bounds
-/// checks match Node and throw `RangeError` on violation:
-///   - `byteOffset % BYTES_PER_ELEMENT == 0` (alignment),
-///   - `byteOffset <= buffer.byteLength`,
-///   - when `length` is omitted, the remaining bytes are a whole multiple of
-///     `BYTES_PER_ELEMENT`,
-///   - `byteOffset + length * BYTES_PER_ELEMENT <= buffer.byteLength`.
-/// `offset_value` / `length_value` are the raw NaN-boxed arguments
-/// (`undefined` when absent) so `ToIndex` runs here. A non-`ArrayBuffer`
-/// source routes to the normal `js_typed_array_new` constructor.
-#[no_mangle]
-pub extern "C" fn js_typed_array_view(
-    kind: i32,
-    source: f64,
-    offset_value: f64,
-    length_value: f64,
-) -> *mut TypedArrayHeader {
-    let kind = kind as u8;
-    let bits = source.to_bits();
-    if (bits >> 48) != 0x7FFD {
-        return js_typed_array_new(kind as i32, source);
-    }
-    let addr = (bits & 0x0000_FFFF_FFFF_FFFF) as usize;
-    if !crate::buffer::is_registered_buffer(addr) || !crate::buffer::is_any_array_buffer(addr) {
-        return js_typed_array_new(kind as i32, source);
-    }
-    let bpe = elem_size_for_kind(kind) as i64;
-    let src = addr as *const crate::buffer::BufferHeader;
-    let total_len = unsafe { (*src).length as i64 };
-
-    let offset = typed_array_view_to_index(offset_value);
-    if bpe > 1 && offset % bpe != 0 {
-        throw_range_error(
-            format!(
-                "start offset of {} should be a multiple of {}",
-                name_for_kind(kind),
-                bpe
-            )
-            .as_bytes(),
-        );
-    }
-    if offset > total_len {
-        throw_range_error(
-            format!("Start offset {offset} is outside the bounds of the buffer").as_bytes(),
-        );
-    }
-
-    let length_jv = crate::value::JSValue::from_bits(length_value.to_bits());
-    let elem_count = if length_jv.is_undefined() {
-        let remaining = total_len - offset;
-        if bpe > 1 && remaining % bpe != 0 {
-            throw_range_error(
-                format!(
-                    "byte length of {} should be a multiple of {}",
-                    name_for_kind(kind),
-                    bpe
-                )
-                .as_bytes(),
-            );
-        }
-        remaining / bpe
-    } else {
-        let requested = typed_array_view_to_index(length_value);
-        if offset + requested * bpe > total_len {
-            throw_range_error(format!("Invalid typed array length: {requested}").as_bytes());
-        }
-        requested
-    };
-
-    // The native byte layout matches between BufferHeader and TypedArrayHeader,
-    // so copy the backing bytes at `offset` directly to preserve element values.
-    let count = elem_count.max(0) as u32;
-    let ta = typed_array_alloc(kind, count);
-    if count > 0 {
-        unsafe {
-            let src_data = crate::buffer::buffer_data(src).add(offset as usize);
-            let dst = data_ptr_mut(ta);
-            ptr::copy_nonoverlapping(src_data, dst, (count as i64 * bpe) as usize);
-        }
-    }
-    ta
 }
 
 /// Copy elements from one typed array into a new typed array of `dst_kind`,
